move constatns to new course cnostants files
